servers

+++ script/lib.launcher.sh
@@ -9,10 +9,12 @@
 : ${xdl_log_dir="${workingdir}/logs"}
 : ${xdl_conf="launcher.sample.ini"}
 
+: ${xdl_api_mode=0}
+
 xdl_home="https://github.com/amentain/launcher.sh"
 xdl_latest_release="https://api.github.com/repos/amentain/launcher.sh/releases/latest"
 xdl_latest_release_cacheTime=$(( 2 * 60 * 60 ))
-xdl_version="0.3.6"
+xdl_version="0.5.0"
 
 xdl_install_path="${BASH_SOURCE}"
 
@@ -43,6 +45,29 @@ function debug {
         done
         eval ${cmd}
     fi
+}
+
+function bytes2human() {
+    # converts a byte count to a human readable format in IEC binary notation (base-1024), rounded to two decimal places for anything larger than a byte. switchable to padded format and base-1000 if desired.
+    local L_BYTES="${1:-0}"
+    local L_PAD="${2:-no}"
+    local L_BASE="${3:-1024}"
+    BYTESTOHUMAN_RESULT=$(awk -v bytes="${L_BYTES}" -v pad="${L_PAD}" -v base="${L_BASE}" 'function human(x, pad, base) {
+         if(base!=1024)base=1000
+         basesuf=(base==1024)?"iB":"B"
+
+         s="BKMGTEPYZ"
+         while (x>=base && length(s)>1)
+               {x/=base; s=substr(s,2)}
+         s=substr(s,1,1)
+
+         xf=(pad=="yes") ? ((s=="B")?"%5d   ":"%8.2f") : ((s=="B")?"%d":"%.2f")
+         s=(s!="B") ? (s basesuf) : ((pad=="no") ? s : ((basesuf=="iB")?(s "  "):(s " ")))
+
+         return sprintf( (xf " %s\n"), x, s)
+      }
+      BEGIN{print human(bytes, pad, base)}')
+    return $?
 }
 
 ###### INI #############################################################################################
@@ -93,7 +118,7 @@ function __findJAR {
 }
 
 function __getPID {
-    echo "/tmp/`echo $1 | tr " " "_"`.pid"
+    echo "/tmp/`echo $1 | tr -d '"' | tr ' ' '_'`.pid"
 }
 
 ###### Daemon start & stop ###############################################################################
@@ -118,7 +143,7 @@ function check_alive {
 function startDaemon_java {
     local DAEMON="$1"
     local JRUN="$2"
-    local DEBUG="${3:0}"
+    local DEBUG=${3:0}
     if [ "xx$DAEMON" = "xx" ]; then
         error "no DAEMON" 1
     fi
@@ -142,7 +167,7 @@ function startDaemon_java {
 
     echo "Starting $DAEMON..."
     local LOG_PREFIX="${xdl_log_dir}/`echo ${DAEMON} | tr " " "_"`"
-    if [ "${DEBUG}" -ne 1 ]; then
+    if [ ${DEBUG} -ne 1 ]; then
         rotateLogs "${LOG_PREFIX}"
 
         nohup java $params $JRUN $args > "${LOG_PREFIX}-output.log" 2> "${LOG_PREFIX}-error.log" &
@@ -157,7 +182,7 @@ function startDaemon_java {
 function startDaemon_node {
     local DAEMON="$1"
     local NODE_FILE="$2"
-    local DEBUG="${3:0}"
+    local DEBUG=${3:0}
     if [ "xx$DAEMON" = "xx" ]; then
         error "no DAEMON" 1
     fi
@@ -191,7 +216,7 @@ function startDaemon_node {
 function startDaemon_plain {
     local DAEMON="$1"
     local DAEMON_FILE="$2"
-    local DEBUG="${3:0}"
+    local DEBUG=${3:0}
     if [ "xx$DAEMON" = "xx" ]; then
         error "no DAEMON" 1
     fi
@@ -209,7 +234,7 @@ function startDaemon_plain {
     fi
 
     echo "Starting $DAEMON..."
-    local LOG_PREFIX="$LOG_DIR/`echo ${DAEMON} | tr " " "_"`"
+    local LOG_PREFIX="${xdl_log_dir}/`echo ${DAEMON} | tr " " "_"`"
     if [ ${DEBUG} -ne 1 ]; then
         rotateLogs "${LOG_PREFIX}"
 
@@ -220,6 +245,47 @@ function startDaemon_plain {
         echo ${DAEMON_FILE} $params $args || return $?
         ${DAEMON_FILE} $params $args || return $?
     fi
+}
+
+function statusDaemon {
+    local DAEMON=`echo $1 | sed 's/"//g'`
+    if [ "xx$DAEMON" = "xx" ]; then
+        error "no DAEMON" 1
+    fi
+
+    if [ "xx${XDL_STATUS_HEADER}" = "xx" ]; then
+        statusHeader
+    fi
+
+    local lenDif=$(( ${XDL_STATUS_MAXLEN} - ${#DAEMON} ))
+    local PID_FILE=`__getPID "$DAEMON"`
+    if check_alive "$PID_FILE"; then
+        local pCPU=$(ps -o "%cpu" -p `cat ${PID_FILE}` | tail -n1)
+        local pMEM=$(ps -o "%mem" -p `cat ${PID_FILE}` | tail -n1)
+        local pRSS=$(ps -o "rss"  -p `cat ${PID_FILE}` | tail -n1)
+
+        bytes2human $((${pRSS} * 1024))
+        printf "%s%${lenDif}s\t%d\t%s\t%s\t%s\n" "${DAEMON}" " " "`cat ${PID_FILE}`" "${pCPU}" "${pMEM}" "${BYTESTOHUMAN_RESULT}"
+    else
+        printf "%s%${lenDif}s\t%s\n" "${DAEMON}" " " "--"
+    fi
+    return 1
+}
+
+function statusHeader {
+    XDL_STATUS_HEADER=yes
+
+    local ttl="DAEMON"
+    local maxLen=`__getMaxDaemonLength`
+    if [ ${maxLen} -lt ${#ttl} ]; then
+        maxLen=${#ttl}
+    fi
+    maxLen=$(( (($maxLen / 2) + 1) * 2 ))
+
+    local lenDif=$(( maxLen - ${#ttl} ))
+    printf "%s%${lenDif}s\t%s\t%s\t%s\t%s\n" "${ttl}" " " "PID" "CPU,%" "MEM,%" "RSS,bytes"
+
+    XDL_STATUS_MAXLEN=${maxLen}
 }
 
 function stopDaemon {
@@ -273,27 +339,37 @@ function rotateLogs {
     local LOG_PREFIX=$1
     local dtNow=$(date +%F_%T)
 
-    mv -f "${LOG_PREFIX}-output.log" "${LOG_PREFIX}-output.${dtNow}.log"
-    mv -f "${LOG_PREFIX}-error.log" "${LOG_PREFIX}-error.${dtNow}.log"
+    if [ -f "${LOG_PREFIX}-output.log" ]; then
+        mv -f "${LOG_PREFIX}-output.log" "${LOG_PREFIX}-output.${dtNow}.log"
+        rm -f `ls ${LOG_PREFIX}-output.*.log | sort -r | sed 1,5d`
+    fi
 
-    rm -f `ls ${LOG_PREFIX}-output.*.log | sort -r | sed 1,5d`
-    rm -f `ls ${LOG_PREFIX}-error.*.log  | sort -r | sed 1,5d`
+    if [ -f "${LOG_PREFIX}-error.log" ]; then
+        mv -f "${LOG_PREFIX}-error.log" "${LOG_PREFIX}-error.${dtNow}.log"
+        rm -f `ls ${LOG_PREFIX}-error.*.log  | sort -r | sed 1,5d`
+    fi
 }
 
 ###### Updated ###########################################################################################
 
-function version_gt() { test "$(echo "$@" | tr " " "\n" | sort -g | tail -n 1)" != "$1"; }
+function version_gt { test "$(echo "$@" | tr " " "\n" | sort -g | tail -n 1)" != "$1"; }
 
 function __getLatestRelease {
     local cache="${xdl_tmp}/release.json"
     local upgrade=$1
+    local force=$2
+
     if [ "${upgrade}" != "1" ]; then
         upgrade=0
     fi
 
+    if [ "${force}" = "force" ]; then
+        rm -f "${cache}"
+    fi
+
     if [ -f "${cache}" ]; then
         local now=`date +%s`
-        local cacheTime=`stat -f '%Sm' -t '%s' "${cache}"`
+        local cacheTime=`__getMTime "${cache}"`
         if [ $(($now - $cacheTime)) -gt ${xdl_latest_release_cacheTime} ]; then
             rm -f "${cache}"
         fi
@@ -301,7 +377,7 @@ function __getLatestRelease {
 
     if ! [ -f "${cache}" ]; then
         wget -q -O "${cache}" "${xdl_latest_release}" || error "Can't update sorry" 1
-        chmod 777 "${cache}" # global writable
+        chmod 666 "${cache}" # global writable
         touch "${cache}" # set current date
     fi
 
@@ -325,6 +401,22 @@ function __getLatestRelease {
         fi
     else
         echo "Already up to date"
+    fi
+}
+
+function __getMTime {
+    local fName="${1}"
+
+    if [ ! -f "${fName}" ]; then
+        echo "Can't read file ${fName}" >&2
+        echo -1
+        exit 1
+    fi
+
+    if [ $(uname) = "FreeBSD" -o $(uname) = "Darwin" ]; then
+        stat -f '%Sm' -t '%s' "${fName}"
+    else
+        date +%s -r "${fName}"
     fi
 }
 
@@ -352,6 +444,24 @@ function __getAvailableDaemons {
     echo ${available_daemons}
 }
 
+function __getMaxDaemonLength {
+    if [ "xx${XDL_MAX_DAEMON_LENGTH}" != "xx" ]; then
+        echo ${XDL_MAX_DAEMON_LENGTH}
+        exit 0
+    fi
+
+    local len=1
+    for d in `__getAvailableDaemons`
+    do
+        local l=${#d}
+        if [ ${len} -lt ${l} ]; then
+            len=${l}
+        fi
+    done
+    XDL_MAX_DAEMON_LENGTH=${len}
+    echo ${len}
+}
+
 function __checkDaemon {
     for d in `__getAvailableDaemons`
     do
@@ -365,6 +475,7 @@ function __checkDaemon {
 function __getDaemonParams() {
     local d=$1
     daemon=${d}
+    eval owner=\${xdml_${d}_owner}
     eval artifact=\${xdml_${d}_artifact}
     eval node_file=\${xdml_${d}_node_file}
     eval daemon_file=\${xdml_${d}_daemon_file}
@@ -413,6 +524,7 @@ function __showUsage {
     echo "Usage:"
     echo "${cmd} start       - to start daemon in background"
     echo "${cmd} stop        - to stop daemon"
+    echo "${cmd} status      - displays daemon status"
     echo "${cmd} restart     - to stop and start daemon"
     echo
     echo "${cmd} run         - to start daemon in the current console"
@@ -430,15 +542,20 @@ function __showUsage {
 function __runDaemonCommand {
     local command_name=$1
     local command_sub_name=$2
+    local run_cmd_1=""
+    local run_cmd_2=""
+    local debug=0
 
     case "$command_name" in
         start|run)
-            debug=0
             if [ "$command_name" == "run" ]; then
                 debug=1
             fi
-            ${runner} "$debug"
-            return $?
+            run_cmd_1="${runner} $debug"
+        ;;
+
+        status)
+            run_cmd_1="statusDaemon \"$daemon\""
         ;;
 
         stop)
@@ -447,7 +564,7 @@ function __runDaemonCommand {
                 force=1
             fi
 
-            stopDaemon "$daemon" "$force"
+            run_cmd_1="stopDaemon \"$daemon\" \"$force\""
         ;;
 
         restart)
@@ -456,92 +573,124 @@ function __runDaemonCommand {
                 force=1
             fi
 
-            stopDaemon "$daemon" "$force"
-            ${runner} "0"
-            return $?
+            run_cmd_1="stopDaemon \"$daemon\" \"$force\""
+            run_cmd_2="${runner} $debug"
         ;;
 
         *)
             __showUsage
+            return 1
         ;;
     esac
+
+    if [ "${command_name}" = "status" -o "xx${owner}" = "xx" -o $(id -un) = "${owner}" ];
+    then
+        ${run_cmd_1}
+        ${run_cmd_2}
+        return $?
+    else
+        printf "Using sudo -u ${owner}...\n"
+        sudo -H -u ${owner} bash <<RUN
+debug="${debug}"
+verbose="${verbose}"
+this_script="${this_script}"
+workingdir="${workingdir}"
+
+xdl_tmp="${xdl_tmp}"
+xdl_log_dir="${xdl_log_dir}"
+xdl_conf="${xdl_conf}"
+
+xdl_api_mode=1
+
+. ${xdl_install_path}
+
+${run_cmd_1}
+${run_cmd_2}
+exit \$?
+RUN
+        return $?
+    fi
 }
-
-# Going home
-cd ${workingdir}
-
-# Testing tmp access
-mkdir -p ${xdl_tmp} || error "TMP is not writable: can't write to ${xdl_tmp}" 2
-chmod 777 ${xdl_tmp} # global writable
-
-# Check & parse ini
-dmList=""
-dmCount=0
-if [ -f "${xdl_conf}" ]; then
-    __ini_get "${xdl_conf}" "xdml_"
-    dmList=`__getAvailableDaemons`
-    dmCount=$(( `echo ${dmList} | wc -w` * 1 ))
-fi
-
-# Check input params
-if [ "xx$1" == "xx" ]; then
-    __showUsage
-fi
 
 # home sweet home
 cd ${workingdir}
 
-first=$1; shift
-case "${first}" in
-    "all")
-        # Check available daemons
-        if [ ${dmCount} -eq 0 ]; then
-            __showUsage
-        fi
+# api or not api?
+if [ ${xdl_api_mode} -ne 1 ]; then
+    # Testing tmp access
+    mkdir -p ${xdl_tmp} || error "TMP is not writable: can't write to ${xdl_tmp}" 2
+    chmod 777 ${xdl_tmp} # global writable
 
-        for d in ${dmList}
-        do
-            __getDaemonParams ${d}
-            __runDaemonCommand $@
-        done
-    ;;
+    # Check & parse ini
+    dmList=""
+    dmCount=0
+    if [ -f "${xdl_conf}" ]; then
+        __ini_get "${xdl_conf}" "xdml_"
+        dmList=`__getAvailableDaemons`
+        dmCount=$(( `echo ${dmList} | wc -w` * 1 ))
+    fi
 
-    "update")
-        __getLatestRelease
-    ;;
+    # Check input params
+    if [ "xx$1" == "xx" ]; then
+        __showUsage
+    fi
 
-    "upgrade")
-        __getLatestRelease 1
-    ;;
-
-    "version")
-        __showVersion
-    ;;
-
-    "install")
-        echo "Not implemented"
-        exit 1
-    ;;
-
-    *)
-        # Check available daemons
-        if [ ${dmCount} -eq 0 ]; then
-            __showUsage
-        fi
-
-        if [ ${dmCount} -eq 1 ];
-        then
-            __getDaemonParams ${dmList}
-            __runDaemonCommand "${first}" $@
-        else
-
-            if __checkDaemon ${first}; then
-                printf "no daemon \"${first}\" found\n\n\n"
+    first=$1; shift
+    case "${first}" in
+        "all")
+            # Check available daemons
+            if [ ${dmCount} -eq 0 ]; then
                 __showUsage
             fi
 
-            __getDaemonParams ${first}
-            __runDaemonCommand $@
-        fi
-    ;;
-esac
+            for d in ${dmList}
+            do
+                __getDaemonParams ${d}
+                __runDaemonCommand $@
+            done
+        ;;
+
+        "update")
+            __getLatestRelease 0 $@
+        ;;
+
+        "upgrade")
+            __getLatestRelease 1 $@
+        ;;
+
+        "version")
+            __showVersion
+        ;;
+
+        "install")
+            echo "Not implemented"
+            exit 1
+        ;;
+
+        *)
+            # Check available daemons
+            if [ ${dmCount} -eq 0 ]; then
+                __showUsage
+            fi
+
+            if [ ${dmCount} -eq 1 ];
+            then
+                __getDaemonParams ${dmList}
+                if [ "${first}" == "${dmList}" ]; then
+                    __runDaemonCommand $@
+                else
+                    __runDaemonCommand "${first}" $@
+                fi
+            else
+
+                if __checkDaemon ${first}; then
+                    printf "no daemon \"${first}\" found\n\n\n"
+                    __showUsage
+                fi
+
+                __getDaemonParams ${first}
+                __runDaemonCommand $@
+            fi
+        ;;
+    esac
+fi
